pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autonomyx-adk"
version = "0.1.0"
description = "Autonomyx ADK — One SDK, all agent frameworks. Auto-provisions identity, auth, tools."
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [{ name = "OpenAutonomyx" }]
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.45.0"]
langchain = ["langchain>=0.3.0", "langchain-anthropic>=0.3.0"]
autogen = ["autogen-agentchat>=0.4.0"]
crewai = ["crewai>=0.80.0"]
google = ["google-adk>=0.1.0"]
all = ["autonomyx-adk[anthropic,langchain,crewai]"]
dev = ["pytest", "pytest-asyncio", "respx"]

[project.scripts]
autonomyx = "autonomyx.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/autonomyx"]
